+++ src/services/administrator.ts
@@ -241,6 +241,31 @@ const AdministratorServices = {
       console.log(error);
     }
   },
+  uploadPatientDiet: async (id: string, token: string, asset: Asset) => {
+    try {
+      const request = new RequestData(
+        Networking.administrator.uploadPatientDiet,
+        token,
+      );
+      request.setParams(`/${id}`);
+
+      let formData = new FormData();
+      const file = {
+        name: `${uuid.v4()}.${uriToFileType(asset.uri ?? '.jpg')}`,
+        type: asset.type,
+        size: asset.fileSize,
+        uri: asset.uri,
+      };
+      formData.append('file', file);
+
+      request.setFormData(formData);
+
+      const response = await request.request();
+      return response;
+    } catch (error) {
+      console.log(error);
+    }
+  },
 };
 
 export default AdministratorServices;